+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Mapbox welcomes participation and contributions from everyone.
 ## Bug fixes 🐞
 * Fix the crash when identifying if device is connected to WiFi.
 
+# 10.12.1
+## Bug fixes 🐞
+* Fix missing data id in `source-data-loaded` event for empty GeoJSON data.
+
+## Dependencies
+* Update gl-native to v10.12.1.
+
 # 10.12.0 March 23, 2023
 ## Features ✨ and improvements 🏁
 * Add `data-id` argument to `GeoJsonSource` data update methods. `data-id` is later attached to the `SourceDataLoadedEventData` event and allows to track the specific `GeoJsonSource` update.